Move SignUpModel.kt to right package
