+++ app/src/main/java/by/alexandr7035/affinidi_id/data/AuthRepository.kt
@@ -1,6 +1,7 @@
 package by.alexandr7035.affinidi_id.data
 
 import by.alexandr7035.affinidi_id.data.model.SignInModel
+import by.alexandr7035.affinidi_id.data.model.SignUpModel
 
 interface AuthRepository {
     suspend fun signUp(
