+++ app/src/main/java/by/alexandr7035/affinidi_id/presentation/registration/AuthViewModel.kt
@@ -1,11 +1,10 @@
 package by.alexandr7035.affinidi_id.presentation.registration
 
-import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import by.alexandr7035.affinidi_id.data.AuthRepository
-import by.alexandr7035.affinidi_id.data.SignUpModel
+import by.alexandr7035.affinidi_id.data.model.SignUpModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
